# docs/public/openapi.json
{
	"components": {
		"schemas": {
...
			"ErrorDetail": {
				"additionalProperties": false,
				"properties": {
					"location": {
						"description": "Where the error occurred, e.g. 'body.items[3].tags' or 'path.thing-id'",
						"type": "string"
					},
					"message": {
						"description": "Error message text",
						"type": "string"
					},
					"value": {
						"description": "The value at the given location"
					}
				},
				"type": "object"
			},
			"ErrorModel": {
				"additionalProperties": false,
				"properties": {
					"$schema": {
						"description": "A URL to the JSON Schema for this object.",
						"examples": ["/api/v1/schemas/ErrorModel.json"],
						"format": "uri",
						"readOnly": true,
						"type": "string"
					},
					"detail": {
						"description": "A human-readable explanation specific to this occurrence of the problem.",
						"examples": ["Property foo is required but is missing."],
						"type": "string"
					},
					"errors": {
						"description": "Optional list of individual error details",
						"items": {
							"$ref": "#/components/schemas/ErrorDetail"
						},
						"type": ["array", "null"]
					},
					"instance": {
						"description": "A URI reference that identifies the specific occurrence of the problem.",
						"examples": ["https://example.com/error-log/abc123"],
						"format": "uri",
						"type": "string"
					},
					"status": {
						"description": "HTTP status code",
						"examples": [400],
						"format": "int64",
						"type": "integer"
					},
					"title": {
						"description": "A short, human-readable summary of the problem type. This value should not change between occurrences of the error.",
						"examples": ["Bad Request"],
						"type": "string"
					},
					"type": {
						"default": "about:blank",
						"description": "A URI reference to human-readable documentation for the error.",
						"examples": ["https://example.com/errors/example"],
						"format": "uri",
						"type": "string"
					}
				},
				"type": "object"
			},
...
			"LoginInputBody": {
				"additionalProperties": false,
				"properties": {
					"$schema": {
						"description": "A URL to the JSON Schema for this object.",
						"examples": ["/api/v1/schemas/LoginInputBody.json"],
						"format": "uri",
						"readOnly": true,
						"type": "string"
					},
					"email": {
						"description": "Email address used to sign in. It is normalized before lookup.",
						"examples": ["user@example.com"],
						"type": "string"
					},
					"password": {
						"description": "Plain-text password to verify. It is never returned by the API.",
						"examples": ["correct-horse-battery-staple"],
						"type": "string",
						"writeOnly": true
					}
				},
				"type": "object"
			},
			"LoginOutputBody": {
				"additionalProperties": false,
				"properties": {
					"$schema": {
						"description": "A URL to the JSON Schema for this object.",
						"examples": ["/api/v1/schemas/LoginOutputBody.json"],
						"format": "uri",
						"readOnly": true,
						"type": "string"
					},
					"user": {
						"$ref": "#/components/schemas/UserResponse",
						"description": "Authenticated user."
					}
				},
				"required": ["user"],
				"type": "object"
			},
...
			"RegisterInputBody": {
				"additionalProperties": false,
				"properties": {
					"$schema": {
						"description": "A URL to the JSON Schema for this object.",
						"examples": ["/api/v1/schemas/RegisterInputBody.json"],
						"format": "uri",
						"readOnly": true,
						"type": "string"
					},
					"displayName": {
						"description": "Name shown in the app.",
						"examples": ["Jane Doe"],
						"type": "string"
					},
					"email": {
						"description": "Email address used to sign in.",
						"examples": ["user@example.com"],
						"type": "string"
					},
					"password": {
						"description": "Plain-text password. It is stored only as an Argon2id hash.",
						"examples": ["correct-horse-battery-staple"],
						"type": "string",
						"writeOnly": true
					}
				},
				"type": "object"
			},
			"RegisterOutputBody": {
				"additionalProperties": false,
				"properties": {
					"$schema": {
						"description": "A URL to the JSON Schema for this object.",
						"examples": ["/api/v1/schemas/RegisterOutputBody.json"],
						"format": "uri",
						"readOnly": true,
						"type": "string"
					},
					"user": {
						"$ref": "#/components/schemas/UserResponse",
						"description": "Created user."
					}
				},
				"required": ["user"],
				"type": "object"
			},
...
			"UserResponse": {
				"additionalProperties": false,
				"properties": {
					"displayName": {
						"description": "Name shown in the app.",
						"examples": ["Jane Doe"],
						"maxLength": 100,
						"type": "string"
					},
					"email": {
						"description": "Normalized email address used to sign in.",
						"examples": ["user@example.com"],
						"format": "email",
						"type": "string"
					},
					"id": {
						"description": "User UUID.",
						"examples": ["11111111-1111-1111-1111-111111111111"],
						"format": "uuid",
						"type": "string"
					}
				},
				"required": ["id", "email"],
				"type": "object"
			}
		},
		"securitySchemes": {
			"probeAuth": {
				"description": "Probe runtime credential header using the format `Probe \u003csecret\u003e`.",
				"in": "header",
				"name": "Authorization",
				"type": "apiKey"
			},
			"sessionCookieAuth": {
				"description": "HTTP-only session cookie set by `/auth/register` and `/auth/login`.",
				"in": "cookie",
				"name": "netstamp_session",
				"type": "apiKey"
			}
		}
	},
	"info": {
		"description": "Controller HTTP API for Netstamp.",
		"title": "Netstamp API",
		"version": "v1"
	},
	"openapi": "3.1.0",
	"paths": {
...
		"/auth/login": {
			"post": {
				"description": "Verify an email and password pair, then set an HTTP-only session cookie for the authenticated user. Invalid credentials always return the same unauthorized response so callers cannot distinguish an unknown email from a password mismatch.",
				"operationId": "loginUser",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginInputBody"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LoginOutputBody"
								}
							}
						},
						"description": "OK"
					},
					"401": {
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorModel"
								}
							}
						},
						"description": "Unauthorized"
					},
					"422": {
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorModel"
								}
							}
						},
						"description": "Unprocessable Entity"
					},
					"500": {
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorModel"
								}
							}
						},
						"description": "Internal Server Error"
					}
				},
				"summary": "Login user",
				"tags": ["Auth"]
			}
		},
...
		"/auth/register": {
			"post": {
				"description": "Create a user account with a normalized email address, display name, and password. On success, sets an HTTP-only session cookie for immediate API access.",
				"operationId": "registerUser",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegisterInputBody"
							}
						}
					},
					"required": true
				},
				"responses": {
					"201": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RegisterOutputBody"
								}
							}
						},
						"description": "Created"
					},
					"409": {
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorModel"
								}
							}
						},
						"description": "Conflict"
					},
					"422": {
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorModel"
								}
							}
						},
						"description": "Unprocessable Entity"
					},
					"500": {
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorModel"
								}
							}
						},
						"description": "Internal Server Error"
					}
				},
				"summary": "Register user",
				"tags": ["Auth"]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}